Minor changes to dictionary switching mechanism
Humble PR (and firsty on GH) for ginny keyboard to try and enable dictionary switching with a dictionary reset on failsafe.
Will update as I try the keyboard a bit more. Apologies for any code formatting errors.

Still struggling to get eyes onto what happens with that first if statement, but I think it is correctly formed to receive the SET_STICKY(0) arguments when sent.